admin/requirements.txt
Flask
pandas
psycopg2-binary
SQLAlchemy
requests